ms/boardManagementService/src/main/java/com/elysium/reddot/ms/board/infrastructure/inbound/rest/processor/CreateBoardProcessor.java
package com.elysium.reddot.ms.board.infrastructure.inbound.rest.processor;

import com.elysium.reddot.ms.board.application.data.dto.ApiResponseDTO;
import com.elysium.reddot.ms.board.application.data.dto.BoardDTO;
import com.elysium.reddot.ms.board.application.service.BoardApplicationServiceImpl;
import lombok.AllArgsConstructor;
import org.apache.camel.Exchange;
import org.apache.camel.Processor;
import org.springframework.http.HttpStatus;
import org.springframework.stereotype.Component;

@Component
@AllArgsConstructor
public class CreateBoardProcessor implements Processor {

    private final BoardApplicationServiceImpl boardService;

    @Override
    public void process(Exchange exchange) {
        BoardDTO boardDTO = exchange.getIn().getBody(BoardDTO.class);
        BoardDTO boardCreated = boardService.createBoard(boardDTO);
        ApiResponseDTO apiResponseDTO = new ApiResponseDTO(HttpStatus.CREATED.value(),
                "Board with name " + boardCreated.getName() + " created successfully", boardCreated);
        exchange.getMessage().setBody(apiResponseDTO);
    }
}