ms/boardManagementService/src/test/java/com/elysium/reddot/ms/board/infrastructure/inbound/rest/processor/CreateBoardProcessorTest.java
package com.elysium.reddot.ms.board.infrastructure.inbound.rest.processor;

import com.elysium.reddot.ms.board.application.data.dto.ApiResponseDTO;
import com.elysium.reddot.ms.board.application.data.dto.BoardDTO;
import com.elysium.reddot.ms.board.application.service.BoardApplicationServiceImpl;
import org.apache.camel.CamelContext;
import org.apache.camel.Exchange;
import org.apache.camel.impl.DefaultCamelContext;
import org.apache.camel.support.DefaultExchange;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.HttpStatus;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class CreateBoardProcessorTest {

    private CreateBoardProcessor createBoardProcessor;

    @Mock
    private BoardApplicationServiceImpl boardService;

    private CamelContext camelContext;

    @BeforeEach
    void setUp() {
        camelContext = new DefaultCamelContext();
        createBoardProcessor = new CreateBoardProcessor(boardService);
    }

    @Test
    @DisplayName("given valid board when createBoard is called then board created successfully")
    void givenValidBoard_whenCreateBoard_thenBoardCreatedSuccessfully() {
        // arrange
        BoardDTO boardDTO = new BoardDTO(null, "name", "Name", "Board description");
        BoardDTO createdBoardDTO = new BoardDTO(1L, "name", "Name", "Board description");

        ApiResponseDTO expectedApiResponse = new ApiResponseDTO(HttpStatus.CREATED.value(),
                "Board with name " + createdBoardDTO.getName() + " created successfully", createdBoardDTO);

        Exchange exchange = new DefaultExchange(camelContext);
        exchange.getIn().setHeader("CamelHttpUri", "/boards");
        exchange.getIn().setBody(boardDTO);

        // mock
        when(boardService.createBoard(boardDTO)).thenReturn(createdBoardDTO);

        // act
        createBoardProcessor.process(exchange);
        ApiResponseDTO actualApiResponse = exchange.getMessage().getBody(ApiResponseDTO.class);

        // assert
        assertEquals(expectedApiResponse.getStatus(), actualApiResponse.getStatus());
        assertEquals(expectedApiResponse.getMessage(), actualApiResponse.getMessage());
        assertEquals(expectedApiResponse.getData(), actualApiResponse.getData());
    }


}
